Renamed `SpotlessTaskBase` to `SpotlessTask`, and the implementation details live in `SpotlessTaskImpl`.
